Move graph walkers module into linters module
