+++ src/databricks/labs/ucx/source_code/linters/graph_walkers.py
