+++ tests/unit/source_code/test_jobs.py
@@ -28,7 +28,7 @@ from databricks.labs.ucx.source_code.graph import (
     DependencyResolver,
 )
 from databricks.labs.ucx.source_code.jobs import JobProblem, WorkflowLinter, WorkflowTaskContainer
-from databricks.labs.ucx.source_code.graph_walkers import LintingWalker
+from databricks.labs.ucx.source_code.linters.graph_walkers import LintingWalker
 from databricks.labs.ucx.source_code.notebooks.loaders import NotebookResolver, NotebookLoader
 from databricks.labs.ucx.source_code.used_table import UsedTablesCrawler
